+++ src/main/java/domain/piece/LinearMove.java
@@ -1,7 +1,7 @@
 package domain.piece;
 
+import domain.JanggiBoard;
 import domain.JanggiCoordinate;
-import domain.board.JanggiBoard;
 
 public interface LinearMove {
     void validateLinearMove(JanggiBoard janggiBoard, JanggiCoordinate from, JanggiCoordinate to);
